Tweaked for more efficiency.
Right, so I removed an if(). Basically, it takes the OR of the two
conditions. So if now > endsong, that evaluates false. BUT, if
playforever == 1 then the whole thing evalutes to true. I like to move
as much logic out of the main loop as I can to minimize needless
processing while the music is playing.